finish StackVec implementation
`StackVec` exposes a `Vec`-like API when given pre-allocated memory. This allows user
to pre-allocate memory and then abstract away it's management to a common data structure.
In a programming language like `C` the usage of the `static` keyword would pre-allocate
data in the static section of a binary or through stack allocations from local varable
declarations.

A number of traits were implemented to get the tests passing: